Generated by Spark: ok this is great. next, we want to add the first zero config feature. when capture is submitted, we will run the first inference. which shuld not tell much. without and cofig, it can infer a due date. and might have a decent idea if the text is a note or a reminder or a action item. but it would have zero context of what collection to put in. so it will have very low confidence for "migrating" so we need to start by adding work flows for processing the item based on confidence. it is my opinion that the first thing that we need to know is what type of item was captured. we are starting the mvp with three possible types. 1. Note, jsut something that needs to be remembered. 2. action item something that the user needs to do. 3. reminder which is something that will be visible to the user at a certain time in the future (this can be an action item or a notw when the reminder time comes up, but for now it is somethig that we need to store and then surface at a certian point in the future.
so the next thing to add is the first decision, decide the type. so build a basic inference engine that can use key words and phrases in the input to infer if it is a reminder (remind me to ..., follow up on...) or an action item (i need to.., Create a..., Call x and....) if it infers a type it will assign a confidence to the inference, and then it will prompt the user with the type it inferred the confidence level and whether it needs to change, or, it will ask what type it is. then it will update the item with the type attribute and update the learning object storage with any data that it can use to ingfer type in the future.